Moved node definitions
