+++ app/lib/client/client.dart
@@ -18,7 +18,7 @@ import 'package:sph_plan/themes.dart';
 
 import '../shared/apps.dart';
 import '../shared/shared_functions.dart';
-import 'fach.dart';
+import '../shared/types/fach.dart';
 
 class SPHclient {
   final statusCodes = {
